renamed module to infoRender
