+++ src/super/i-block/modules/info-render/interface.ts
